requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-docx
pdf2image
pillow
